Changed packet transfer method (now nodes are transferred).

+++ TCPDumpNetworkVisualizer/src/com/uh/nwvz/client/components/InformationPopup.java
@@ -23,7 +23,7 @@ public class InformationPopup extends PopupPanel {
     }
 	
 	public void setData(String host, int packetsSent, int packetsReceived, 
-			int kbSent, int kbReceived) {
+			long kbSent, long kbReceived) {
 		
 		infoTable.setText(0, 1, host);
 		infoTable.setText(1, 1, String.valueOf(packetsSent));
